renamed queries.py to utils.py, deleted few traces left of pypika
