+++ parsec/backend/postgresql/realm_queries/maintenance.py
@@ -16,10 +16,10 @@ from parsec.backend.realm import (
     RealmNotInMaintenanceError,
 )
 from parsec.backend.postgresql.handler import send_signal
-from parsec.backend.postgresql.utils import query
 from parsec.backend.postgresql.message import send_message
-from parsec.backend.postgresql.queries import (
+from parsec.backend.postgresql.utils import (
     Q,
+    query,
     q_organization_internal_id,
     q_user,
     q_device_internal_id,
