+++ parsec/backend/postgresql/user_queries/device_invitation.py
@@ -9,9 +9,9 @@ from parsec.backend.user import (
     DeviceInvitation,
 )
 from parsec.backend.postgresql.handler import send_signal
-from parsec.backend.postgresql.utils import query
-from parsec.backend.postgresql.queries import (
+from parsec.backend.postgresql.utils import (
     Q,
+    query,
     q_organization_internal_id,
     q_device,
     q_device_internal_id,
