+++ parsec/backend/postgresql/vlob_queries/create.py
@@ -6,9 +6,9 @@ from triopg import UniqueViolationError
 
 from parsec.api.protocol import DeviceID, OrganizationID
 from parsec.backend.vlob import VlobAlreadyExistsError
-from parsec.backend.postgresql.utils import query
-from parsec.backend.postgresql.queries import (
+from parsec.backend.postgresql.utils import (
     Q,
+    query,
     q_organization_internal_id,
     q_realm_internal_id,
     q_device_internal_id,
